Reverted to 4.1.0

+++ src/Media/Geometry/Dimensions.php
@@ -25,8 +25,7 @@ class Dimensions
     {
         $this->_width = (int) $width;
         $this->_height = (int) $height;
-        // NOTE: MUST `float`-cast to FORCE float even when dividing EQUAL ints.
-        $this->_aspectRatio = (float) ($this->_width / $this->_height);
+        $this->_aspectRatio = $this->_width / $this->_height;
     }
 
     /**
@@ -52,7 +51,7 @@ class Dimensions
     /**
      * Get stored aspect ratio for these dimensions.
      *
-     * @return float
+     * @return int
      */
     public function getAspectRatio()
     {
