+++ src/Media/MediaDetails.php
@@ -37,8 +37,7 @@ abstract class MediaDetails
      */
     public function getAspectRatio()
     {
-        // NOTE: MUST `float`-cast to FORCE float even when dividing EQUAL ints.
-        return (float) ($this->getWidth() / $this->getHeight());
+        return $this->getWidth() / $this->getHeight();
     }
 
     /**
